初步完成报告并融合stride,添加一些debug报告
初步完成报告，并实现stride比较，但并未shijiyingy

尝试融合stride，但是有bug

修复bug
成功将不怕溢出的stride加入进代码

+++ os/src/config.rs
@@ -26,5 +26,5 @@ pub const MEMORY_END: usize = 0x88000000;
 /// The base address of control registers in Virtio_Block device
 pub const MMIO: &[(usize, usize)] = &[(0x10001000, 0x1000)];
 ///lab5 StrideInfo
-pub const INIT_PRIORITY:isize = 16;
-pub const BIGSTRDE:isize = isize::MAX >> 20; 
+pub const INIT_PRIORITY:u64 = 16;
+pub const BIGSTRDE:u64 = u64::MAX >> 20;
